refactor: split main into smaller files
